Add Logger object
This object is intended to be a fast and versatile way to write
information to disk. It allocated a buffer at start up where log entries
are written to. The buffer is flushed when full or on timeout.

The StringBuilder was augmented to evaluate formats, so the Logger
simply forwards the evaluation to it. Instead of relying on type-unsafe
variadic functions, the variadic.h/.c files were introduce which
implement the V() macro, a way to convert arrays of etherogeneous values
to an array of packed tagged unions.